requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
websockets==12.0
opencv-python==4.9.0.80
numpy==1.26.4
mediapipe==0.10.9
scikit-learn==1.4.1.post1
joblib==1.3.2
python-dotenv==1.0.1
gunicorn==21.2.0 